fix: tarifa de servicio cobra un 5% del monto final y ahora todo es bigDecimal

+++ backend/ms-pedido/src/main/java/com/seminario/ms_pedido/model/DetalleCarrito.java
@@ -17,10 +17,10 @@ public class DetalleCarrito {
     private String idItem;
     private String productoId;
     private Integer cantidad;
-    private Double montoUnitario;
+    private BigDecimal montoUnitario;
     private String observaciones;
 
-    public DetalleCarrito(String productoId, Integer cantidad, Double montoUnitario, String observaciones) {
+    public DetalleCarrito(String productoId, Integer cantidad, BigDecimal montoUnitario, String observaciones) {
         this.idItem = UUID.randomUUID().toString(); 
         this.productoId = productoId;
         this.cantidad = cantidad;
